[simd] Fix lookup of interface functions.
We need to use qualified lookup here, because we are in `namespace
Dune::Simd::Overloads`, and the functions we want to use are in `namespace
Dune::Simd`.  `Overloads` contains functions of the same name as the ones we
want to use, so unqualified lookup stops looking there and does not try
looking in `Simd`.  Since then the signatures don't match the arguments,
everything fails...