[simd] Do not test masks for operator++.
`operator++` (either kind) on bools is deprecated by the standard, so we
probably should not start requiring it for masks, even if the current Vc masks
support it.  Not even checking for it has the benefit of avoid a deprecation
warning by clang, that cannot easily be shut up via pragmas (there isn't
really a good place to put the pragma...).