add a new class Dune::copy_ptr
this new class behaves somewhat similar to a unique_ptr, in the sense
that it guarantees single ownership. In addition it offers a well
defined way to clone/duplicate the object inside the unique_ptr.
Internally it uses type-erasure for deleter and creator, which is in
particular necessary if the pointer is only a pointer to an abstract
base class.