Add specialized code for inverting 3x3 dense matrix.
I think this specialization was there for inverting a FieldMatrix in the
early DUNE days. Having this improved performance in a downstream module
when using ILU0.

The code is taking from FMatrixHelp::invertMatrix and adapted to the dense
matrix interface. This code used by an ancient (pre-git, pre-module)
geometry implementation.