[cmake] Add dune_execute_process with better error checking
CMake by default does not check the return codes of any
commands invoked through execute_process. This can lead
to very undesirable behaviour. Manually checking the return
code is tedious, so this commit introduces a dune-prefixed
version, which does the job.

This will be used by the python build system integration soon
to be migrated to dune-common.