Remove (index|integer)_sequence fallback to fix #73
All supported compiler/standard library combinations have it (so that
this change does no harm)
Libc++, however, does not declare __cpp_lib_integer_sequence, causing
the fail to check and the fallback to be used. This leads to problems
because code is already freely mixing Dune::Std::*_sequence and
std::* expecting them to be the same, which with Libc++ they are not.