[builder] use CMake for dependency tracking
Instead of generating one CMake target "generated_module", we now define a
CMake target for every module create in dune-py. Before loading the module,
we always call cmake to make sure the module is rebuilt if dependencies change.

Note: Changes in the source code of the generated module remain untracked for
      now.