[dune-py] use version number to identify structural changes in the dune-py module
When the structure of dune-py changes, we simply bump the version number
(revision) of the dune-py module. Upon a release, we will update the major and
minor number and reset the revision number.

If the version of an existing dune-py module does not equal the version we would
generate ourselves, we issue a runtime error, asking the user to remove the
dune-py directory.