[shared_ptr] Fix casting version of stackobject_to_shared_ptr
The version of `stackobject_to_shared_ptr` that takes two type
arguments and casts the returned shared_ptr had the wrong order of
type arguments and was in fact never called. Instead, by explicitly
specifying the template parameter, you upcasted the passed-in
reference to the specified type when calling the function. As the
deleter doesn't do anything, this did not crash the program later on.

This fixes the problem, but requires some trickery to keep itself out
of the overload set if the user does **not** want to cast.