[SIMD] Fixup after filter-branch.
The previous commit is the result of a filter-branch operation of the
feature/proper-simd-interface branch that was needed because the commit hook
complains about trailing whitespace in markdown files with certain names, even
though certain forms of trailing whitespace are part of the markdown language.
There was an attempt to fix the filter (see !336), which poeple did not like.
Since it is unclear whether there is a merge filter that will reject a merge
to master, a filter-branch operation is needed to retroactively change the
name of the file from README.md to DESIGN.md, which the filter does not check.

This commit adjusts the name of the file in the doxygen documentation, too.

The filter-branch command was something like

  git filter-branch --tree-filter "mv dune/common/simd/README.md dune/common/simd/DESIGN.md" --msg-filter "sed 's/README/DESIGN/g'" base..tmp

where

  base=46fc88c6ae82c1da28cc3e4418745d55d62f8eb4
  tmp=8f89f0b9f5356bff7ce807b1f5b7bdd70a1a7617