[Vc] Introduce a proxy around Vc lane accesses.
This is needed because newer versions of Vc introduce their own proxies as the
result of expression like `v[0]`.  These proxies cannot be constructed outside
of Vc, not even by copying or moving, so we cannot pass them out of our
`lane()` function.  We work around this by introducing our own proxy class,
that we can copy.

That proxy class was originally introduced for !193 (Proper SIMD Interface)
and has been backported for the 2.6 Dune release.

Closes: #59