[SIMD][Vc] Include the new simd/vc.hh from old simd.hh
This is needed for transition: E.g. `DenseMatrix` has been switched to use the
new interface, but the solvers in ISTL have not.  They still include
`<dune/common/simd.hh>`, expecting that this is enough to make `FieldMatrix`
work with Vc.  But `FieldMatrix` now requires that whoever hands it Vc types
includes `<dune/common/simd/vc.hh>`.  By doing that from
`<dune/common/simd.hh>` we ensure that this requirement is met.