[filledArray()] Kill superfluous construct to make constexpr well-defined.
A template function may be marked `constexpr` as long as there could be a
combination of template and runtime arguments where it, indeed, satisfies the
requirements for a constexpr function (and only for those arguments will a
call actually be a constant expression).

The extra template parameter `Array` was intended to exploit that loophole: in
theory a user could provide a different type there that behaved like a
`std::array`, except that it would indeed be `constexpr` even in older
standards when `std::array` wasn't.  This construction turns out to be
unnessary, because the template parameter `T` already exploits that loophole:
There could be some user-defined type `T` for which the user could provide an
overload of `std::array` that has, indeed, `constexpr` semantics.

As a bonus, that allows us to spell out the return type of `filledArray()` so
that it will show up in Doxygen.