[cmake][dune_add_test] Overwrite command and cmd_args if a cmake guard is false
If a complicated user defined command was specified, it didn't lead to the dummy
executable (that return 77) being run but the user defined command instead. This
patch makes sure that the correct command is executed which then marks the test
as skipped and not as failed. If a test is a parallel and tested with multiple
number of ranks all these test are shown as skipped.