[simd/looptest] Generate instantiations.
Currently, the `std::complex` instantiations fail because `lane(0,
constant_vec) = something;` is well-formed.  I'm assuming this is due to
`lane()` returning a `std::complex` prvalue, and since `std::complex` is a
class the assignment is rewritten to `lane(0,
constant_vec).operator=(something)`.  This works because you can call
non-const member functions on class prvalues.  Unless the prvalue is declared
`const` -- which we don't want to do because the inhibits moving from the
prvalue.

I'm gonna have to relax the test, I guess.