[simd/test.hh] Allow vector-scalar testing without vector-vector operations.
Previously the vector-scalar tests assumed that a vector-vector operation was
available and would use that to check the result.  This fails for certain
vectorclass types.

This is a bigger rewrite because I could not understand my own code from a few
months ago.  Hopefully the documentation is better this time...