Use pkg_resources.working_set instead of pip internal
In recent pip versions (>= 9), the pip.get_installed_distributions()
method was moved to pip._internal. It is strongly suggested to not use
this method, see e.g. https://github.com/pypa/pip/issues/5243

It is recommended to use pkg_resources.working set to inquire for any
information on installed distributions. This commit ports the extraction
of the version number of a package to this interface.