DuneMPI: append flags to `COMPILE_OPTIONS` instead of `COMPILE_FLAGS`
`COMPILE_FLAGS` is a string, but `set_property(... APPEND ...)`
appends to a list.  If the list of compile flags is not empty this
results in a superfluous semicolon in the compile flags which will
break the compiler invocation.

For example, when `MPI_DUNE_COMPILE_FLAGS=-lpthread`, calling
`add_dune_mpi_flags` twice for the same target would result in the
compiler command containing `-lpthread;-lpthread`.

This patch makes use of the `COMPILE_OPTIONS` property instead which
is a list.