[SIMD] Only try and apply assignment to non-const lvalues in tests.
A non-const class rvalue is, in fact, assignable, unless you take special
care.  In the tests we try to apply each operator to `Scalar<V>` in an
unevaluated context, and if that succeeds we require that it works for `V`
analogously to how it does for `Scalar<V>`.

That heuristic fails if `Scalar<V>` is a class (such as `std::complex`),
because assignment works even to rvalues (usually), but trying to apply the
assignment to a `V` rvalue and then capturing and examining the result will
lead to undefined behaviour.