[simd/test.hh] Don't test proxies for scalar operation support.
In the tests for operations we check whether an operation is supported on
scalars, and if so, we also test is in simd types.  This suppresses tests for
e.g. `%` on simd types made up of floating point types.  Before this patch we
sometimes did the check on proxies instead of scalars.  This could lead to
tests be incorrectly not suppressed.