Work around a nasty compiler bug in GCC 6.3
GCC 6.3 errors out with an ICE or simply hangs when instantiating a certain variant of operator^= on
BitSetVectorReference. This patch works around the problem by forcing the actual XOR instruction
into a separate function for that compiler. The workaround is probably horribly slow, but it
shouldn't have any impact on unaffected compilers.

This probably needs backporting.