Try to install Python virtualenv with pip before forcing get-pip onto users
Way back when we first integrated a virtualenv into the Dune build process, there was a nasty bug in
several Linux distributions that made it impossible to get a standard virtualenv with pip installed
inside of it, so we decided to completely skip the default installation of pip into the virtualenv
and instead download get-pip to install it. This is not always appropriate, so now the flag
DUNE_PYTHON_ALLOW_GET_PIP has to be enabled for it, but without it enabled, we still force users to
manually install pip into the virtualenv.

Meanwhile, pretty much all distributions managed to get their act together (even Ubuntu 14.04 now
correctly installs pip into the virtualenv if the required packages are installed). In order to make
life easier for our users, this patch changes the default behavior: It now tries to create a
virtualenv with pip installed first, but if that fails, it reverts to the old behavior.