[bugfix] fix memory fault upon module unloading in Python 3.6
In Python 3.6 we observe memory faults at program termination. These
result from the static variables caching the Python object wrapping the
type registry: Upon termination, the type registry seems to be destroyed
even if references to it remain.

Unfortunately, it is not clear whether this is a problem within
Pybind11. Patching to the latest bugfix release did not help. For now,
we implement the simplest way around the problem: Do not cache the type
registry, but refetch the pointer each time the type registry is used.
While this has a negative impact on performance, it should be negligible
to subsequent CMake calls.