Let cmake disable search for modules with pkg-config.
For whatever reason we still allow to search for a module
with pkg-config if CMake was not successful. This leads to
the weired situation that if we pass
-DCMAKE_DISABLE_FIND_PACKAGE_dune-uggrid=ON to cmake and
dune-uggrid is installed in a default sytem path we will end
up with a CMakeCache.txt actually using dune-uggrid.

This patch is the minimal change to stop searching for a
DUNE module with pkg-config if CMake was instructed not to
use it.