[bugfix] Fix default constructor of IntegralRangeInterator
In c++-14 all value-initilized forward iterators
should behave like the end-iterator of a hypothetic
empty container and furthermore should all compare equal.
Here, this would lead to undefined behaviour since
the `value_` member is not initialized.

To avoid this we no initilize it by 0 which mimics
'past end of empty range'