bin/duneproject: Expand tabs and fix indentation
As a shell script, duneproject is not covered by the scripts preventing tabs
in indentation (or simply was exempt from the Great Reformatting).  Also, the
tab width was inconsistent.

This expands all tabs and uses a consistent indentation of 2.  It also
instructs both emacs and vi to use these settings, since their default
indentation for shell scripts seems to differ.