[tools] Make dune-ctest more robust when parsing CTest output
As it turns out, older versions of CTest do not output a "Completion Status" for
tests that were not run, so we need a fallback. The only place for information
is the test output, which we parse for certain key phrases to detect tests that
were skipped or had missing requirements (typically due to compilation
failures).