[Simd][Test] Implement rejecting of rebound types
This adds an additional template parameter `RebindAccept` to `check()`.  This
is a type predicate defaulting to `Std::to_true_type`, that can be used to
reject types resulting from rebinding.  A rejection triggers a static
assertion in the check.

This can be used to ensure missing explicit instantiations for do not go
unnoticed.