[bugfix] Fix Std::optional assignment operator
You cannot conditionally enable assigment operators,
because templated `operator=` does not count as user defined
assignment operator. As a remedy you could do this via a conditional
base class. But I think it's not worth to spend the effor to gain
the extra bit of compatibility. In the end we did not implement
all the required conditions for `operator=` anyway.