[cmake] Fix detection of ParMETIS v4 in cmake
ParMETIS has changed the casing of its function names in the past. In version 3, the all lowercase version was supported alongside the newer camel case version. In version 4, the lower case version vanished. The cmake test should check for the new function instead of the old. All prior testing with version 4 (which we DO support) seems to date back to the autotools time. The problem is clearly visible in dune-istl CI builds, where ParMETIS is considered to be not found in [Ubuntu 18.04 builds](https://gitlab.dune-project.org/core/dune-istl/-/jobs/86757), while it is found in [Debian 9 builds](https://gitlab.dune-project.org/core/dune-istl/-/jobs/86754). Thanks to @santiago.ospina and @dmitry.mazilkin for reporting the failure.