Move method 'power' to namespace Dune::Common
The recent commit b6eca40ec1db52b1ec9b8bef29524e1066f40bc8
introduced a method called 'power' in the namespace Dune.
Rather quickly, this lead to name clashes with the method of the
same name in dune-functions.  The method there is contained
in the namespace Dune::Functions::BasisFactory, which many
people import completely.

As a remedy, this patch moves the 'power' method into the new
namespace Dune::Common.