Make DenseMatrix robust wrt to proxy matrices
For a proxy-matrix you cannot simply copy the derived type `MAT`
to get a self-contained copy. This will just copy the proxy and
the new proxy will point to the same value. To solve such issues
we have the `AutonomousValue` mechanism which provides (by
spezialization) a type which allows to store a self-contained copy.

To make this work we need to call `luDecomposition` of
`DenseMatrix<AutonomousValue<MAT>>` instead of `DenseMatrix<MAT>`.
It turnes out that, surprisingly, this method is non-static which
is also fixed by this commit.