[className()] Make demangling helper `inline`.
This is needed to avoid several instances of the helper conflicting with each
other during linking.

Alternatively, the helper could be put into the library (but it is kind of
small for that).  Or it could be made `static`, though I'm unsure whether it's
OK to call an internal linkage function from an inline function (`className()`
is implicitly inline as it is a template).