Don't pass the cpu flags detected by Vc to the compiler
Vc detects CPU flags based on the name of the CPU.  This can be a problem in
virtualization environment, when the name of the CPU is visible, but not all
features are enabled in the guest.

Also, enabling all detected features in the the compiler makes it difficult to
impossible for the user to force his own feature flags, e.g. to cross-build
generic executables on hosts with many features to be run later on hosts with
fewer features.

Closes: docker/ci#9