Implement two-argument vector / scalar multiplication
This patch works, but it does not use the appropriate
type propogation rules: Multiplying a vector with a scalar
will always result in a vector using the same number type
as the input vector.

Fixing this on the DenseVector level is not easy, because
it requires a rebinding mechanism, which DenseVector
does not have.