Implement matrix-matrix multiplication for square matrices
I don't think you can currently implement the non-square case
on the DenseMatrix level, because the operator needs to create
a result matrix type with the correct number of rows and
colums.  This is not possibly unless the factors are square,
in which case the result has the same number of rows and columns
as the factors.