[latex][bugfix][win32][cmake] ignore CMAKE_FIND_ROOT_PATH for latexmkrc.cmake
When cross compiling for windows it is necessary to tweak
CMAKE_FIND_ROOT_PATH_MODE_* to modify default behavior of FIND_XXX()
commands to search for headers/libs in the target environment and
search for programs in the build host environment.

This lead to a problem when searching for latexmkrc.cmake.

We now explicitly change the UseLatexMk find_file call to ignore the
CMAKE_FIND_ROOT_PATH and thus beeing able to find the file in the
source directory.