[test][cmake] Fix test command for skipped test
This fixes the bug (#183) introduced with a225b2da.
The command can be a complex path while for skipped tests
we actually want the test command to just be equal to
the target name (executable). This name is
expanded to a full path using the TARGET_FILE generator
expression so it should work on all platforms that CMake
supports.