Do not import std::shared_ptr into namespace Dune
This is a left-over from the days where Dune had its own
implementation of std::shared_ptr.  Now that that is long gone,
it violates the principle of least surprise to find shared_ptr
in Dune::

As a side effect, this patch removes most of the code from
sharedptrtest.cc.  As we don't have our own shared_ptr implementation
anymore, there is no more reason to test it.