Get rid off fortran compiler usage for Lapack/Blas.
FindLAPACK and FindBLAS check for the name mangling with added
underline as a fallback, anyway. Hence there is no need for using
a Fortran compiler. With this commit we now simply add another check
for a LAPACK function to see whether it has an underline added and
adapt the existing macro in config.h accordingly.

Closes #184