Fix warnings, add comments in fvectorconversion1d.cc test
This is a regression test, the threw up undefined variable warnings.  unsure
about the correct fix, I tried to understand the test.  In addition to fixing
the warnings, I left some comments too, in the hope future readers will have a
little easier time understanding what it is supposed to do.

The warnings showed up in
https://gitlab.dune-project.org/core/dune-common/-/jobs/139437#L665 (debian:10
clang-7-libcpp-17):
```
[ 60%] Building CXX object dune/common/test/CMakeFiles/fvectorconversion1d.dir/fvectorconversion1d.cc.o
/builds/core/dune-common/dune/common/test/fvectorconversion1d.cc:86:16: warning: unused variable 'mfv' [-Wunused-variable]
      MiddleFV mfv;
               ^
/builds/core/dune-common/dune/common/test/fvectorconversion1d.cc:107:16: warning: unused variable 'mfv' [-Wunused-variable]
      MiddleFV mfv;
               ^
2 warnings generated.
```

See also: !735, !562