[cmake][mpi] Use correct guard MPI
The dune CMake configuration actually need `MPI_C_FOUND` to be true. If this is true `HAVE_MPI` is set. This is used with this commit.

According to the CMake 3.1 doc `MPI_FOUND` is actually discouraged. It is later reintroduced but is only true if MPI has been found for all requested languages. Which is also not what we want.