Make eigenvector computation of diagonal 2x2 matrices more stable
The eigenvector code used to fail for certain seemingly simple
diagonal 2x2 matrices.  The reason was that the code tried
to avoid certain all-zero matrix columns by testing whether
the norm is (close to) zero.  That proved to be surprisingly
unstable.  The fix here is simple: Pick the column with the
larger column norm (we know they are not both zero).

Also add a test case to make sure the bug is gone.