[cmake] allow user/module specific ignore files for cpack
the CPACK_SOURCE_IGNORE_FILES is hard wired in DuneMacros.cmake. While
this is a good idea to ensure a working setup in most cases, it blocks
module specific ignores.

We add an additional list CPACK_SOURCE_IGNORE_FILES_EXTRA that can be
specified by the user and will by appended to the
CPACK_SOURCE_IGNORE_FILES list.