[python] Add support for --module=mod to bin/setup-dunepy.py
This adds an optional parameter `--module=mod` to `setup-dunepy.py`.
If this parameter is passed, only `mod` and its dependencies are
used as dependencies of the newly created dune-py module. This allows
to do
```
dunecontrol --opts=my.opts --module=foo all
setup-dunepy.py --opts=my.opts --module=foo install
```
which would otherwise fail if `setup-dunepy.py` tries to process
found modules that have not been build by `dunecontrol`. If the
option is not passed all found modules are used as dependencies
as before.