[python] store cmake argumets simply as a list
cmake arguments were stored as a dict, having mainly definitions in mind.

This lead to several issues, as cmake has many different kinds of
commmand line arguments.

Instead of storing these arguments as a dict, we now only store them as
a list, which can directly be passed to Popen, when spawning the cmake
process.

Fixes #238.