Add a script that build a PyPI mirror and pushes it to the dune-common package registry
Such a mirror can be used for running CI in isolated environments.
The script only needs to be run once by a core developer whenever
Python dependencies of Dune modules change or new Python versions
are released.

Mirror PyPI for all Dune packages that are available on PyPI

Base the PyPI mirror script on pip-download

added a few more packages

[fix]