all dune modules are searched at the beginning now possibly before dune-py has been constructed. The __init__.py file in dune.grid uses 'haveAssert' to check if 'AlbertaGrid' is available. This checks 'config.h' in dune-py which caused an error. The fix here simply return 'False' in case dune-py/config.h is not available.
correctly use the tagfile '.noconfigure' to make sure dune-py is rebuild
each time a new dune module is installed.
To avoid a 'recursive include' issue moved the generation of dune-py based on the
'template' to the 'Builder' (where it was before).

Also fixed a bug (which was partially a feature avoiding recursive includes up until now):
there was a script dune/generator/builder.py which contained the 'Builder' class and
in dune/generator/__init__.py we setup a variable 'builder' which then shadowed the script.
This did then cause problems in one place. Renaming the 'builder.py' script then demonstrated
the issue with the recursive import.

use *.template for dune-py template files which should be copied without
the ending, e.g., CMakeLists.txt (now CMakeLists.txt.template)
Also renamed dune.module to dune.module.template to be certain that it will
not be picked up by mistake

add cmake install command for the cmake_metadata files and the files in the
dune-py template to make sure that skbuild adds them correctly to the wheel

check for matplotlib in dune.plotting

Partial fix for export of build dirs for all deps modules to metadat file.
In previous version only first builddir was written.
Issue with this version: space is used to separate the directories.
Also make dune-py depend on modules without their own python package if another module
(with bindings) depends on it.

minor fixes