[Python] Always write cmake metadata files to dune/data/<package>.cmake.
The advantage is that we don't need to load all dune subpackages anymore to extract the metadata and can so avoid the issues with circular dependencies.
Instead is it enough to import the new submodule dune.data.
The CMAKE_METADATA_FILE parameter to dune_python_install_package is not needed anymore and is ignroed.