Constrain dot to cases where it makes sense
This constraints the non-vector overload of `Dune::dot(a,b)`
to cases where the first argument is a number in the sense
of `Dune::IsNumber`.
This helps to avoid ambiguity if one wants to implement both,
`dot(a,b)` and `a*b` manually for a custom type. Without this
patch, the availability of `a*b` activates `Dune::dot(a,b)`
concurrently to the custom implementation.