[python] lock cmake in dune-py module to avoid cmake raise condition
Running python scripts in parallel might trigger different make calls in
parallel. If the cmake setup of dnue-py changed in the meantime this
triggers parallel cmake reconfiguration runs. Parallel cmake calls are
not allowed and as a consequence the complete build may fail. We take
care of such parallel cmake calls by additional locking in the dune-py
CMakeLists.txt