Improve specializations of Hybrid::size()
The interfaces tested subsequently for `Hybrid::size(const T& t)` are now

* standard tuple-like interface `std::tuple_size<T>::value` (e.g. needed for `std::tuple`),
* static constexpr `T::size()` (e.g. needed for `Dune::TupleVector`),
* dynamic member function `t.size()` (e.g. needed for `std::vector`).

This patch removes a special case for `FieldVector` which is already covered
by the `static constexpr size()` case. To ensure this, a corresponding test
case was added. Also the documentation of the different cases was fixed.

Notice that this does not work necessarily with matrices, because many of them
only provide a (potentially `static constexpr`) method `N()` instead of `size()`.